Create Hero component
