+++ src/components/Layout.tsx
@@ -8,7 +8,7 @@ import {
 import { Outlet } from "react-router-dom";
 import Footer from "./Footer/Footer";
 import Nav from "./Navbar/Nav";
-import theme from "./Theme/theme";
+import theme from "../theme/theme";
 
 const Layout = () => {
   return (
